requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astroid==3.3.8
attrs==25.3.0
blobfile==3.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dill==0.3.9
distro==1.9.0
Django==5.2.1
djangorestframework==3.16.0
et_xmlfile==2.0.0
filelock==3.18.0
fire==0.7.0
fonttools==4.55.8
fsspec==2025.3.2
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
isort==5.13.2
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
llama_stack==0.2.2
llama_stack_client==0.2.2
lxml==5.3.0
mando==0.7.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mccabe==0.7.0
mdurl==0.1.2
numpy==2.2.2
openai==1.76.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.51
pyaml==25.1.0
pycryptodomex==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pylint==3.3.3
pyparsing==3.2.1
PySide6==6.8.1.1
PySide6_Addons==6.8.1.1
PySide6_Essentials==6.8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
radon==6.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
setuptools==79.0.1
shiboken6==6.8.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlparse==0.5.3
termcolor==3.0.1
tiktoken==0.9.0
tomlkit==0.13.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
wcwidth==0.2.13
XlsxWriter==3.2.1
